+++ RigidBodies/PhysicsProject/Planet.cpp
@@ -15,7 +15,7 @@ Planet::~Planet()
 }
 
 //--------------------------------------------------------------------------------
-void Planet::Initialize(std::string name, float radius, double mass, Vector3D position, Vector3D velocity)
+void Planet::Initialize(std::string name, float radius, float mass, Vector3D position, Vector3D velocity)
 {
 	PhysicObject::Initialize(mass, position, velocity);
 	m_Name = name;
@@ -33,7 +33,7 @@ void Planet::Initialize(std::string name, float radius, double mass, Vector3D po
 	m_DebugText.push_back(m_Name);
 
 	m_DrawPosition = Vector3D(position);
-	m_DrawPosition *= 1.0e-7;
+	m_DrawPosition *= 1.0e-7f;
 
 }
 
@@ -46,7 +46,7 @@ void Planet::Update(float dt)
 //--------------------------------------------------------------------------------
 void Planet::Draw()
 {
-	m_DrawPosition = m_Position * 1.0e-8;
+	m_DrawPosition = m_Position * 1.0e-8f;
 
 	glPushMatrix();
 
@@ -72,7 +72,7 @@ void Planet::DrawDebugText()
 	m_DebugText[2] = "Position: " + m_Position.ToString();
 
 
-	for (int i = 0; i < m_DebugText.size(); i++)
+	for (unsigned int i = 0; i < m_DebugText.size(); i++)
 	{
 		glRasterPos3f(m_DrawPosition.X - m_Velocity.X, m_DrawPosition.Y + m_Radius + (i / 10.0f), m_DrawPosition.Z + m_Velocity.Z);
